fee: mandatory 2.5% dev fee to the fork address
Sets this fork's dev fee to 2.5% (was 1% upstream) and points it at the
fork's own address. Documented in README.md.

  kDevAddress    btx1zrjgx4cnkqc097k27tl57utuxx8vm6fc4w8n3xf7kxntwc9phvp8strh3ef
  kDevFeeMinPct  2.5   (90 seconds of every 3600)

The fee gate InDevFeeWindow() took an int percent, so 2.5 was not
representable; widened to double. The three config paths (json / env / cli)
also parsed the fee as int, which truncated 2.5 to 2, and applied no floor --
so --dev-fee 1 or 0 stuck. All three now go through ClampDevFee(), which
floors at kDevFeeMinPct, caps at 100, and treats NaN and unparseable input as
the floor. The pool path floors again at read time because it can run before
the main() clamp depending on config ordering.

Also removed a dead 'cfg.devfee = 1;' immediately after the main() clamp that
would have silently reset the fee after logging the clamped value.

devfee_window_test rewritten to anchor on 2.5%: it sweeps a full period one
second at a time and asserts exactly 90 of 3600 seconds fall in the window,
plus fractional-percent boundaries. Verified in both release artefacts.

No change to the proof-of-work maths; byte-exact gate still passes
(5b1bff3c835b1c8e7816a2cccb181eb2fc30a99d97a971d73108c52a8238acd4).

+++ clean-stack/miner/devfee_window.h
@@ -18,10 +18,10 @@
 // so every restart lands inside the leading dev slice -- which billed 4 of 5 solo blocks to the
 // dev address on a restart-heavy night (2026-08-12) under a 1% fee. The argument keeps its old
 // name only because the arithmetic is identical for any monotonic seconds source.
-inline bool InDevFeeWindow(double uptime_sec, double period_sec, int devfee_pct)
+inline bool InDevFeeWindow(double uptime_sec, double period_sec, double devfee_pct)
 {
-    if (devfee_pct <= 0 || period_sec <= 0.0) return false;
-    const double window = period_sec * (static_cast<double>(devfee_pct) / 100.0);
+    if (devfee_pct <= 0.0 || period_sec <= 0.0) return false;
+    const double window = period_sec * (devfee_pct / 100.0);
     // Position within the current period: uptime_sec mod period_sec (truncated, matches the miner).
     const double pos = uptime_sec - period_sec *
         static_cast<double>(static_cast<uint64_t>(uptime_sec / period_sec));
